Renamed task_timeline to job_timeline
