+++ crates/hyperqueue/src/dashboard/ui/screens/worker/worker_tasks_table.rs
@@ -1,5 +1,5 @@
 use crate::common::format::human_duration;
-use crate::dashboard::data::task_timeline::{DashboardTaskState, TaskInfo};
+use crate::dashboard::data::job_timeline::{DashboardTaskState, TaskInfo};
 use crate::dashboard::ui::styles::table_style_selected;
 use crate::dashboard::ui::terminal::DashboardFrame;
 use crate::dashboard::ui::widgets::table::{StatefulTable, TableColumnHeaders};
